requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.34
psycopg[binary]==3.*

passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.6.1

bcrypt==3.2.2
python-jose[cryptography]==3.3.0
